hotel_booking_system/hotel_ai_agent_service/requirements.txt
fastapi
uvicorn
python-dotenv
langgraph
langchain-openai
httpx